Initial commit
A Kubernetes app that keeps a service internal cluster ip in Kubernetes
synced with a DNS record in a Google Cloud DNS zone.

This can be useful in the case where you are connected to the cluster
network via VPN and still want to address the services via a public
DNS server.