Override DNS name using service annotation
Annotations are the chosen method of configuring Kubernetes
integrations, as described for example here:
https://kubernetes.io/docs/concepts/services-networking/service/#ssl-support-on-aws

This change allows specifying which short hostname the service will get
in the DNS record with the 'external_dns_hostname' annotation. If it is
not set, the service name is used. Domain name is appended in both
cases.

This change is beneficial when deploying services packaged by helm,
since the service name always ends up hyphenated and we would perhaps
like a user friendly hostname instead.