Don't make needless DNS updates
A kubernetes watch update event does not mean that anything has actually
changed. Proper monitoring of kube changes is to check the resource
version. When a manifest change happens, the service version is updated.

As such, to avoid needlessly spamming Google's DNS API endpoint, a guard
is added to ensure we only update the DNS record when a service's
resource version has changed.